project/requirements/dev.txt
-r base.txt

ipython==4.2.0
django-debug-toolbar==1.4
nose==1.3.7
django-nose==1.4.3
codecov==2.0.3
coverage==4.0.3
